refactor action engine to cluster resource manager
the action engine interface is very vague. to help give some more
context on what the interface is for, the name has been changed in
this commit to cluster resource manager.

additionally, a refactor of rds.go and s3.go in the aws package
have been refactored to include the 'manager_' prefix in the file
name.

verification:
- run 'make test/unit'
- ensure tests pass successfully